Move dependencies utility functions to `dependencies.ts`.
